+++ labbook_BE/app/models/Patient.py
@@ -290,13 +290,20 @@ class Patient:
         try:
             cursor = DB.cursor()
 
+            if 'pat_email' not in params or params['pat_email'] is None:
+                params['pat_email'] = ''
+
+            if 'pat_lite' not in params or params['pat_lite'] is None:
+                params['pat_lite'] = 0
+
             cursor.execute('update sigl_03_data '
                            'set id_owner=%(id_owner)s, anonyme=%(anonyme)s, code=%(code)s, code_patient=%(code_patient)s, '
                            'nom=%(nom)s, prenom=%(prenom)s, ddn=%(ddn)s, sexe=%(sexe)s, adresse=%(adresse)s, '
-                           'cp=%(cp)s, ville=%(ville)s, tel=%(tel)s, pat_phone2=%(phone2)s, profession=%(profession)s, '
-                           'nom_jf=%(nom_jf)s, quartier=%(quartier)s, bp=%(bp)s, ddn_approx=%(ddn_approx)s, age=%(age)s, '
-                           'unite=%(unite)s, pat_midname=%(midname)s, pat_nation=%(nationality)s, '
-                           'pat_resident=%(resident)s, pat_blood_group=%(blood_group)s, pat_blood_rhesus=%(blood_rhesus)s '
+                           'cp=%(cp)s, ville=%(ville)s, pat_email=%(pat_email)s, tel=%(tel)s, pat_phone2=%(phone2)s, '
+                           'profession=%(profession)s, nom_jf=%(nom_jf)s, quartier=%(quartier)s, bp=%(bp)s, '
+                           'ddn_approx=%(ddn_approx)s, age=%(age)s, unite=%(unite)s, pat_midname=%(midname)s, '
+                           'pat_nation=%(nationality)s, pat_resident=%(resident)s, pat_blood_group=%(blood_group)s, '
+                           'pat_blood_rhesus=%(blood_rhesus)s, pat_lite=%(pat_lite)s '
                            'where id_data=%(id)s', params)
 
             Patient.log.info(Logs.fileline())
@@ -311,16 +318,22 @@ class Patient:
         try:
             cursor = DB.cursor()
 
+            if 'pat_email' not in params or params['pat_email'] is None:
+                params['pat_email'] = ''
+
+            if 'pat_lite' not in params or params['pat_lite'] is None:
+                params['pat_lite'] = 0
+
             cursor.execute('insert into sigl_03_data '
                            '(id_owner, anonyme, code, code_patient, nom, prenom, ddn, sexe, adresse, cp, ville, '
-                           'tel, pat_phone2, profession, nom_jf, quartier, bp, ddn_approx, age, '
-                           'unite, pat_midname, pat_nation, pat_resident, pat_blood_group, pat_blood_rhesus) '
+                           'pat_email, tel, pat_phone2, profession, nom_jf, quartier, bp, ddn_approx, age, '
+                           'unite, pat_midname, pat_nation, pat_resident, pat_blood_group, pat_blood_rhesus, pat_lite) '
                            'values '
                            '(%(id_owner)s, %(anonyme)s, %(code)s, %(code_patient)s, %(nom)s, %(prenom)s, %(ddn)s, '
-                           '%(sexe)s, %(adresse)s, %(cp)s, %(ville)s, %(tel)s, %(phone2)s, '
+                           '%(sexe)s, %(adresse)s, %(cp)s, %(ville)s, %(pat_email)s, %(tel)s, %(phone2)s, '
                            '%(profession)s, %(nom_jf)s, %(quartier)s, %(bp)s, %(ddn_approx)s, %(age)s, '
                            '%(unite)s, '
-                           '%(midname)s, %(nationality)s, %(resident)s, %(blood_group)s, %(blood_rhesus)s )', params)
+                           '%(midname)s, %(nationality)s, %(resident)s, %(blood_group)s, %(blood_rhesus)s, %(pat_lite)s)', params)
 
             Patient.log.info(Logs.fileline())
 
@@ -418,7 +431,7 @@ class Patient:
                'pat.ddn_approx as birth_approx, pat.age, d_age_unit.label as age_unit, d_sex.label as sex, '
                'pat.pat_midname as middle_name, pat.nom_jf as maiden_name, nat.nat_name as nation, '
                'nat.nat_code as nat_code, pat.pat_resident as resident, pat.cp as zipcode, pat.ville as city, '
-               'pat.profession, d_blood.label as blood_group, d_rhesus.label as blood_rhesus '
+               'pat.profession, d_blood.label as blood_group, d_rhesus.label as blood_rhesus, pat.pat_lite '
                'from sigl_03_data as pat '
                'left join sigl_dico_data as d_sex on d_sex.id_data=pat.sexe '
                'left join sigl_dico_data as d_blood on d_blood.id_data=pat.pat_blood_group '